Rename kern to core
