+++ pkg/core/difficulty.go
@@ -1,4 +1,4 @@
-package kern
+package core
 
 import (
 	"fmt"
