+++ pkg/core/hasht_test.go
@@ -1,10 +1,10 @@
-package kern_test
+package core_test
 
 import (
 	"math/big"
 	"testing"
 
-	. "github.com/levilutz/basiccoin/pkg/kern"
+	. "github.com/levilutz/basiccoin/pkg/core"
 	"github.com/levilutz/basiccoin/pkg/util"
 )
