+++ pkg/core/transaction.go
@@ -1,4 +1,4 @@
-package kern
+package core
 
 // Reference to supposedly-unspent transaction output.
 type Utxo struct {
