+++ pkg/core/transaction_test.go
@@ -1,4 +1,4 @@
-package kern_test
+package core_test
 
 import (
 	"bytes"
@@ -6,7 +6,7 @@ import (
 	"encoding/json"
 	"testing"
 
-	. "github.com/levilutz/basiccoin/pkg/kern"
+	. "github.com/levilutz/basiccoin/pkg/core"
 	"github.com/levilutz/basiccoin/pkg/util"
 )
